requirements.txt
asgiref==3.8.1
Django==5.0.6
django-filter==24.2
pillow==10.3.0
python-dotenv==1.0.1
sqlparse==0.5.0
